prediction_script_1/requirements.txt
joblib==1.3.2
networkx==3.2
pyintergraph==1.3.3
python_igraph==0.11.4
scikit_learn==1.4.1.post1
xgboost==2.0.3
